rename: create domain package
