+++ src/social-service/src/main/java/org/pallete/easelsocialservice/persistence/PaintRepository.java
@@ -1,5 +1,6 @@
 package org.pallete.easelsocialservice.persistence;
 
+import org.pallete.easelsocialservice.persistence.domain.Paint;
 import org.springframework.data.neo4j.repository.Neo4jRepository;
 
 import java.util.Optional;
