+++ src/social-service/src/main/java/org/pallete/easelsocialservice/persistence/domain/Contains.java
@@ -1,18 +1,20 @@
-package org.pallete.easelsocialservice.persistence;
+package org.pallete.easelsocialservice.persistence.domain;
 
+import lombok.Getter;
 import org.springframework.data.neo4j.core.schema.RelationshipId;
 import org.springframework.data.neo4j.core.schema.RelationshipProperties;
 import org.springframework.data.neo4j.core.schema.TargetNode;
 
+@Getter
 @RelationshipProperties
-public class Uses {
+public class Contains {
     @RelationshipId
     private Long id;
 
     @TargetNode
-    private Media media;
+    private Link link;
 
-    public Uses(Media media) {
-        this.media = media;
+    public Contains(Link link) {
+        this.link = link;
     }
 }
