+++ src/social-service/src/main/java/org/pallete/easelsocialservice/persistence/domain/Hashtag.java
@@ -1,4 +1,4 @@
-package org.pallete.easelsocialservice.persistence;
+package org.pallete.easelsocialservice.persistence.domain;
 
 import org.springframework.data.neo4j.core.schema.Id;
 import org.springframework.data.neo4j.core.schema.Node;
