+++ src/social-service/src/main/java/org/pallete/easelsocialservice/persistence/domain/Paint.java
@@ -1,7 +1,5 @@
-package org.pallete.easelsocialservice.persistence;
+package org.pallete.easelsocialservice.persistence.domain;
 
-import lombok.Getter;
-import lombok.Setter;
 import org.springframework.data.neo4j.core.schema.GeneratedValue;
 import org.springframework.data.neo4j.core.schema.Id;
 import org.springframework.data.neo4j.core.schema.Node;
